+++ packages/zero-pg/src/mod.ts
@@ -1,9 +1,12 @@
 export {makeSchemaQuery} from './query.ts';
 export type {Row} from '../../zql/src/mutate/custom.ts';
-export {ZQLPGDatabase, ZQLPGTransaction} from './zql-pg-database.ts';
+export {
+  PGZQLDatabase as ZQLPGDatabase,
+  PGZQLTransaction as ZQLPGTransaction,
+} from './pg-zql-database.ts';
 export {type Connection, type ConnectionTransaction} from './connection.ts';
 export {
-  PostgresConnection,
+  PostgresConnection as PostgresConnection,
   type PostgresLibTransaction,
   type PostgresLibSQL,
 } from './postgres-connection.ts';
